+++ clippy.toml
@@ -1,0 +1,5 @@
+# Environment
+msrv = "1.67"
+
+# Documentation
+missing-docs-in-crate-items = true
