+++ crates/suites/src/lib.rs
@@ -1,1 +1,3 @@
+#![deny(missing_docs)]
 
+//! A collection of mapping suites that implement the mapping-algorithms-rs crate.
